Refactor related folders into boards features
